ADGROUPNAME_Sync_Success.txt
Removed Timestamp from title to simplify logging.
With the sync failure email notification functionality added, it now sends the failed log if a failure occurs, that being the case it is not necessary to hold on to all the success and failure logs. Instead, they overwrite themselves now. Excluding the failure emails, they are a running log of each failure that occurs per group.